13 lesson 52 minutes

+++ DevQuestions/src/DevQuestions.Contracts/Questions/Responses/QuestionResponse.cs
@@ -4,4 +4,4 @@ namespace DevQuestions.Contracts.Questions.Responses;
 
 public record QuestionResponse(
     IEnumerable<QuestionDto> Questions,
-    int TotalCount);
+    long TotalCount);
